Renamed sim.py to hx.py
